[dev] Move PrimaryAction out of button
