+++ src/components/button/Button.stories.ts
@@ -1,6 +1,6 @@
 import { action } from '@storybook/addon-actions';
 import { boolean, select, text } from '@storybook/addon-knobs';
-import { PrimaryAction } from './PrimaryAction';
+import { PrimaryAction } from '../../actions/PrimaryAction';
 import Vue, { PropType } from 'vue';
 import WvuiButton from './Button.vue';
 
@@ -14,21 +14,21 @@ export const configurable = (): Vue.Component =>
 		components: { WvuiButton },
 		props: {
 			enabled: { type: Boolean, default: boolean( 'Enabled', true ) },
-			buttonAction: {
-				type: Object as PropType<keyof typeof PrimaryAction>,
+			action: {
+				type: String as PropType<keyof typeof PrimaryAction>,
 				default: select( 'Action', Object.keys( PrimaryAction ), 'Default' )
 			},
 			quiet: { type: Boolean, default: boolean( 'Quiet', false ) },
 			slotProp: { type: String, default: text( 'Slot', 'Label' ) }
 		},
 		data() {
-			return { ButtonAction: PrimaryAction };
+			return { PrimaryAction };
 		},
 		methods: { click: action( 'click' ) },
 		template: `
 		<wvui-button
 			:disabled="!enabled"
-			:action="ButtonAction[ buttonAction ]"
+			:action="PrimaryAction[ action ]"
 			:quiet="quiet"
 			@click="click"
 		>
@@ -42,7 +42,7 @@ export const combinations = (): Vue.Component =>
 		components: { WvuiButton },
 		data() {
 			return {
-				ButtonAction: PrimaryAction,
+				PrimaryAction,
 				combinations: [
 					{ enabled: true, quiet: false },
 					{ enabled: true, quiet: true },
@@ -64,12 +64,12 @@ export const combinations = (): Vue.Component =>
 				</tr>
 			</thead>
 			<tbody>
-				<tr v-for="action in Object.keys( ButtonAction )" :key="action">
+				<tr v-for="action in Object.keys( PrimaryAction )" :key="action">
 					<th scope="row">{{action}}</th>
 					<td v-for="combo, index in combinations" :key="index">
 						<wvui-button
 							:disabled="!combo.enabled"
-							:action="ButtonAction[ action ]"
+							:action="PrimaryAction[ action ]"
 							:quiet="combo.quiet"
 							@click="click"
 						>
